+++ python_modules/dagster-graphql/dagster_graphql/cli.py
@@ -18,6 +18,7 @@ from dagster.utils.log import get_stack_trace_array
 
 from .client.query import (
     EXECUTE_PLAN_MUTATION,
+    LAUNCH_PIPELINE_EXECUTION_MUTATION,
     START_PIPELINE_EXECUTION_MUTATION,
     START_SCHEDULED_EXECUTION_MUTATION,
 )
@@ -135,6 +136,7 @@ PREDEFINED_QUERIES = {
     'startPipelineExecution': START_PIPELINE_EXECUTION_MUTATION,
     'startScheduledExecution': START_SCHEDULED_EXECUTION_MUTATION,
     'executePlan': EXECUTE_PLAN_MUTATION,
+    'launchPipelineExecution': LAUNCH_PIPELINE_EXECUTION_MUTATION,
 }