+++ python_modules/dagster/dagster/core/definitions/environment_configs.py
@@ -3,9 +3,14 @@ from collections import namedtuple
 from dagster.core.definitions import SolidHandle
 from dagster.core.errors import DagsterInvalidDefinitionError
 from dagster.core.types import Field, List, NamedDict, NamedSelector
-from dagster.core.types.config import ALL_CONFIG_BUILTINS, ConfigType, ConfigTypeAttributes
+from dagster.core.types.config import (
+    ALL_CONFIG_BUILTINS,
+    ConfigType,
+    ConfigTypeAttributes,
+    ConfigTypeKind,
+)
 from dagster.core.types.field import check_opt_field_param
-from dagster.core.types.field_utils import _ConfigComposite
+from dagster.core.types.field_utils import _ConfigHasFields
 from dagster.core.types.iterate_types import iterate_config_types
 from dagster.core.types.runtime import construct_runtime_type_dictionary
 from dagster.utils import camelcase, check, ensure_single_item
@@ -23,7 +28,7 @@ def SystemNamedDict(name, fields, description=None):
     return NamedDict(name, fields, description, ConfigTypeAttributes(is_system_config=True))
 
 
-class _SolidContainerConfigDict(_ConfigComposite):
+class _SolidContainerConfigDict(_ConfigHasFields):
     def __init__(self, name, fields, description=None, handle=None, child_solids_config_field=None):
         self._handle = check.opt_inst_param(handle, 'handle', SolidHandle)
         self._child_solids_config_field = check.opt_inst_param(
@@ -33,6 +38,7 @@ class _SolidContainerConfigDict(_ConfigComposite):
         super(_SolidContainerConfigDict, self).__init__(
             key=name,
             name=name,
+            kind=ConfigTypeKind.DICT,
             fields=fields,
             description=description,
             type_attributes=ConfigTypeAttributes(is_system_config=True),