+++ python_modules/dagster/dagster/core/storage/runs/sqlite/sqlite_run_storage.py
@@ -3,9 +3,10 @@ import os
 from dagster import check
 from dagster.core.definitions.environment_configs import SystemNamedDict
 from dagster.core.serdes import ConfigurableClass, ConfigurableClassData
-from dagster.core.storage.runs import RunStorageSQLMetadata, SQLRunStorage, create_engine
 from dagster.core.types import Field, String
 from dagster.utils import mkdir_p
+
+from ..runs import RunStorageSQLMetadata, SQLRunStorage, create_engine
 
 
 class SqliteRunStorage(SQLRunStorage, ConfigurableClass):