+++ python_modules/libraries/dagster-postgres/dagster_postgres/run_storage.py
@@ -1,7 +1,7 @@
 from dagster import check
 from dagster.core.definitions.environment_configs import SystemNamedDict
 from dagster.core.serdes import ConfigurableClass, ConfigurableClassData
-from dagster.core.storage.runs import RunStorageSQLMetadata, SQLRunStorage, create_engine
+from dagster.core.storage.runs.sqlite import RunStorageSQLMetadata, SQLRunStorage, create_engine
 from dagster.core.types import Field, String